requirements.txt
pyspark==3.5.1
pyyaml>=6.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
lightgbm>=4.6.0
mlflow>=2.22.4
pyarrow>=14.0
boto3>=1.34
great-expectations>=0.18.0
scipy>=1.11
